feat: refactor ScheduleNameForm to use Bloc for state management
